change navigateModifier file
